+++ zipapps/activate_zipapps.py
@@ -9,10 +9,14 @@ def activate(path=None):
     path_str = path.absolute().as_posix()
     if zipfile.is_zipfile(path_str):
         try:
-            _tmp = zipimport.zipimporter(path_str).load_module("ensure_zipapps")
-            modules.pop(_tmp.__name__, None)
-            del _tmp
-            return True
+            spec = zipimport.zipimporter(path_str).find_spec("ensure_zipapps")
+            if spec and spec.loader:
+                _tmp = spec.loader.load_module("ensure_zipapps")
+                modules.pop(_tmp.__name__, None)
+                del _tmp
+                return True
+            else:
+                raise ImportError(path_str)
         except ImportError as err:
             stderr.write(f"WARNING: activate failed for {err!r}\n")
             raise err